requirements.txt
fastapi==0.111.0
opencv-python==4.5.5.62
mediapipe==0.10.21
uvicorn==0.22.0
Pillow==9.4.0
tensorflow==2.12.0
tensorflow-hub==0.13.0
numpy==1.23.5
gunicorn==20.1.0
python-multipart==0.0.9